corrigindo a precisão do cálculo utilizando double

+++ 10-DEBUGGING/app-debbuging/src/edu/walber/debbuging/CalculadoraDeMedias.java
@@ -7,16 +7,16 @@ public class CalculadoraDeMedias {
         Scanner scan = new Scanner(System.in);
         String[] alunos = {"Camila","Lucas","Bruna","Pedro"};
 
-        int media = calculaMediaDaTurma(alunos, scan);
+        double media = calculaMediaDaTurma(alunos, scan);
 
-        System.out.printf("Média da turma %d", media);
+        System.out.printf("Média da turma %.1f", media);
     }
 
-    public static int calculaMediaDaTurma(String[] alunos, Scanner scanner){
-        int soma = 0;
+    public static double calculaMediaDaTurma(String[] alunos, Scanner scanner){
+        double soma = 0;
         for (String aluno : alunos) {
             System.out.printf("Nota do aluno %s: ", aluno);
-            int nota = scanner.nextInt();
+            double nota = scanner.nextInt();
             soma += nota;
         }
         return soma / alunos.length;
